ZEN-21001: Allow to proceed devices with duplicated engineID
Net-snmp library stores device data (engineID, engineTime, engineBoots) in
cache to reuse it. This cause errors when a few devices have the same
engineID.
This commit add a possibility to clean net-snmp cache after the each
snmp call, so devices with the same engineID will proceed properly.

P.S. - please note, that SNMPv3 RFC deny to use a few devices with the
same engineID in a single administrative domain, so this feature violates
RFC